refactor(hub-common): move utils and rename interfaces
affects: @esri/hub-common
